Consecutive tests now work with child_process 11ty run
